+++ get_zscores.c
@@ -83,7 +83,7 @@ te_get_zscores (const EPI *epi, const char *zscores_file,
     /* Read entire file into memory */
     if (-1 == (fd = open (zscores_file, 0)) ||
         0 >= (size = lseek (fd, 0L, 2)) ||
-        NULL == (trec_zscores_buf = malloc ((unsigned) size+2)) ||
+        NULL == (trec_zscores_buf = malloc ((size_t) size+2)) ||
         -1 == lseek (fd, 0L, 0) ||
         size != read (fd, trec_zscores_buf, size) ||
 	-1 == close (fd)) {
